AmigaOS4: the transfer is running at the speed of the poll timer
The measurement names the culprit. 9031917 bytes in 8944 packets over 527642 ms:
1009 bytes each, so full sized RFCOMM frames, arriving seventeen times a second
- one every 59 ms, with a 50 ms poll timer behind it. Round numbers like that do
not come from radios.

The run loop has the USB port's signal in its wait mask and should wake the
moment a transfer completes; this timer was put there as a safety net. It has
been carrying the whole load, and every fix aimed above it - bearer, Single
Response Mode, four reads in flight, sniff mode - was working on something that
was never the constraint. That is why none of them moved the number by a single
kilobyte.

Five milliseconds for now. It is the wrong fix and I would rather say so than
present it as one: the signal should be waking us and finding out why it does
not is the real work. But a tenfold change in throughput cannot be mistaken for
noise, and if it does not appear then this diagnosis is wrong too and the timer
was never the pacer.

Also fixes the rate this prints. Nine megabytes times a thousand overflows 32
bits, so the transfer above reported 837 bytes per second when it managed 17116.

+++ port/amigaos4-libusb/bt_opp_server.c
@@ -149,7 +149,9 @@ static void opp_file_close(bool keep){
                     (unsigned long) opp_received,
                     (unsigned long) opp_packets,
                     (unsigned long) elapsed_ms,
-                    (unsigned long) ((opp_received * 1000) / elapsed_ms),
+                    /* in 64 bits: nine megabytes times a thousand is well past
+                     * what 32 bits hold, and the rate printed was nonsense */
+                    (unsigned long) (((uint64_t) opp_received * 1000) / elapsed_ms),
                     (unsigned long) (opp_packets ? (opp_received / opp_packets) : 0));
     } else {
         remove(opp_path);
